+++ resources/views/default/form/element/hidden.blade.php
@@ -1,3 +1,3 @@
-@if ($displayed)
+@if ($visibled)
   <input {!! $attributes !!} value="{{ $value }}" />
 @endif