+++ resources/views/default/form/element/upload.blade.php
@@ -1,4 +1,4 @@
-@if ($displayed)
+@if ($visibled)
   <div class="form-group form-element-upload {{ $errors->has($name) ? 'has-error' : '' }} well">
     <label for="{{ $name }}" class="control-label {{ $required ? 'required' : '' }}">
       {{ $label }}