+++ src/Contracts/Form/FormElementInterface.php
@@ -24,11 +24,6 @@ interface FormElementInterface extends Renderable, Arrayable, Initializable, Wit
     /**
      * @return bool
      */
-    public function isDisplayed();
-
-    /**
-     * @return bool
-     */
     public function isValueSkipped();
 
     /**